requirements.txt
langchain
langchain-google-genai
langchain-chroma
pandas
python-dotenv
 